move expression.Sorter to new sort package
# Conflicts:
#	memory/table.go
#	sql/expression/function/aggregation/group_concat.go
#	sql/expression/function/aggregation/window_functions.go
#	sql/expression/function/aggregation/window_partition.go
#	sql/iters/rel_iters.go
#	sql/iters/top_rows_iters.go
